requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
asyncpg==0.30.0
attrs==24.3.0
beautifulsoup4==4.12.3
black==24.10.0
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
loguru==0.7.3
Mako==1.3.8
MarkupSafe==3.0.2
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
propcache==0.2.1
psycopg2-binary==2.9.10
pyee==12.0.0
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.6
SQLAlchemy==2.0.37
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
